pdf_word_bot/requirements.txt
python-telegram-bot==20.3
PyPDF2
python-docx